refactor: folder structure changes
