+++ src/portafolio/components/index.js
@@ -1,0 +1,7 @@
+export * from './AboutMe/AboutMe';
+export * from './Contact/Contact';
+export * from './MyProjects/ProjectCard';
+export * from './MyProjects/Projects';
+export * from './Navbar/Navbar';
+export * from './Skills/ItemsSkill';
+export * from './Skills/Skills';
