+++ src/portafolio/pages/MainPage.jsx
@@ -1,16 +1,11 @@
 
 import React from 'react'
-import { AboutMe } from '../components/AboutMe'
-import { Contact } from '../components/Contact'
-import { Navbar } from '../components/Navbar'
-import { Projects } from '../components/Projects'
-import { Skills } from '../components/Skills'
+import { Navbar, AboutMe, Skills, Projects, Contact  } from '../components/index'
 
 export const MainPage = () => {
   return (
     <>
     
-
         <Navbar/>
 
         <AboutMe/>
